requirements.txt
Pillow
tdqm
matplotlib
numpy
pandas
torch
torchvision
scikit-image